tried to stabilize things (no luck)
sometimes you have to set the device-name and/or(?) the resolution of the device
before starting the transfer (e.g. turning rendering on), else it will simply fail
i believe this is a bug in (my old version of) unicap.
e.g. my PS3-eye which definitely only supports 640x480, enumerates a single
format (good) of 768x576 (oops!)



git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@3640 49b2a0ca-c541-4d49-a7ef-d785e4371512