define 'HAVE_LIBFTGL' rather than 'FTGL'
ftgl.h uses 'namespace FTGL', which interferes with an FTGL-define.
i have no idea why it works when doing a "#define FTGL" in GemConfig.h, but does
not work when defining FTGL via property pages.
what i *do* know is that we should rename the 'FTGL' define to 'HAVE_FTGL'