store unicode-representations on the fly
SDL only gives us the unicode of a key on keypress.
since (almost) every keypress is preceded by a keyrelease,
we store the unicode on keypress and can later
retrieve it.
the 'almost' case: if a (world)key is pressed (for the
first time) outside the window, then the focus is shifted
to the window and then the key is released, we don't
know the unicode of the key and thus have to output
some fallback (e.g. 'World_64')
let's see how bad this is in practice