symbolic keys utility
for performance reasons we use integer keys.
but we provide a string2int map, where we can register new keys when
"GemState::getKey("foo") is called the first time, it will return a new key.
next time it get's called it will return the same key.
thus we only need to share the symbolic name, and can use that to initialize a
constant integer key.