put .obj into subdir to prevent nameclashes
there are two files named "rectangle" ("Gem/Rectangle.cpp" and "Geos/rectangle.cpp") and
both would create a file Release/rectangle.obj.
in order to prevent this name-clash, we create Release/Gem/Rectangle.obj for the former.
with VS2010 we should be able to use %(RelativeDir) to create a directory structure for
all .obj files, but with VS2008 we have to do that manually for each file.

confirm http://stackoverflow.com/question/14364362/