pix_data: normalize by (dimen-1)
rather than by (dimen), and then clamp to (dimen-1).
this makes it somewhat clearer what is happening.


the downside is, that this will read the neighbouring pixel
(as compared to the old code), if the input was normalized by dimen.
if the input was normalized by (dimen-1), the output will be the same, except for the
last pixel.