camera: revamped so it now works
- put in into a gemhead to set the camera position for the downstream objects
- just bang it, to get a list suitable for feeding into [gemwin]'s "view" message

this breaks compat with how it used to work, where the global (singleton)
[gemwin] was modified directly.
however, the legacy code has been broken for a while (now that we have multiple windows),
so the break shouldn't be too drastic.

on the plus side, we can now have a "local" camera but also control the
"global" (per gemwin) viewpoint.